[Parser] Fix to work with Lightyear git-tip
First, some signatures:

  execParserT : Monad m => ParserT str m a -> State str -> m (Reply str a)
  runStateT : StateT stateType m a -> stateType -> m (a, stateType)
  evalState : State stateType a -> stateType -> a

Since the signature of execParserT seems to have changed from the time
when Protobuf was last working, we use runStateT instead of evalState,
and pattern-match accordingly.

While at it, modernize some `return` calls to read as `pure`. Finally,
update the tests; the changes are pretty minor, but we take the
opportunity to use a relatively new feature of the Idris language,
triple-quoted string literals.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>